Handle snapshot that is just geth/chaindata

+++ op-geth/docker-entrypoint.sh
@@ -75,9 +75,13 @@ if [ -n "${SNAPSHOT}" ] && [ ! -d "/var/lib/op-geth/geth/" ]; then
     if [ -n "$__found_path" ]; then
       __geth_dir=$(dirname "$__found_path")
       __geth_dir=${__geth_dir%/chaindata}
-      echo "Found a geth directory at ${__geth_dir}, moving it."
-      mv "$__geth_dir" "$__base_dir"
-      rm -rf "$__geth_dir"
+      if [ "${__geth_dir}" = "geth" ];
+         echo "Snapshot extracted into ${__geth_dir}/chaindata"
+      else 
+        echo "Found a geth directory at ${__geth_dir}, moving it."
+        mv "$__geth_dir" "$__base_dir"
+        rm -rf "$__geth_dir"
+      fi
     fi
   fi
   if [[ ! -d /var/lib/op-geth/geth/chaindata ]]; then
